% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_to_rep_symbol.R
\name{control_to_rep_symbol}
\alias{control_to_rep_symbol}
\title{Convert control bytes to Unicode replacement symbol}
\usage{
control_to_rep_symbol(enc_check_results)
}
\value{
A list similar to the one returned by calling
\code{check_column_encoding} but with control bytes sequences converted to
a sequence of Unicode replacement symbols.
}
\description{
For usage by replace_bytes() and find_max_num_seqs().
}

