% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_max_num_seqs.R
\name{find_max_num_seqs}
\alias{find_max_num_seqs}
\title{Assistance in constructing replacement matrices.}
\usage{
find_max_num_seqs(enc_check_results)
}
\arguments{
\item{enc_check_results}{A list returned by calling
\code{check_column_encoding}.}
}
\value{
A named list whose element names correspond to those of
 check_column_encoding or control_to_rep_symbol(check_column_encoding()).
 Each element of this list contains a single number, which indicates the
 number of columns needed for the replacement matrix associated with a column
 name fed to replace_bytes().
}
\description{
Find maximum number of sequences in an element of the results of
 control_to_rep_symbol(check_column_encoding()). Results indicate the number
 of columns needed in the replacement matrix for the respective column name
 fed to replace_bytes(). Note that this function calls
 control_to_rep_symbol() on the results of check_column_encoding(). No action
 is needed by the user for this step.
}

