% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixEncoding.R
\docType{package}
\name{FixEncoding}
\alias{FixEncoding}
\alias{FixEncoding-package}
\title{FixEncoding: A package for locating and replacing invalid UTF-8 bytes}
\description{
The FixEncoding package provides three functions: check_column_encoding.R
for locating non-UTF-8 bytes, control_replace for replacing control bytes,
and diamond_replace.R for replacing invalid UTF-8 bytes.
}

